+++ config.py
@@ -1,0 +1,3 @@
+import os
+
+LIB_DIR = os.path.dirname(os.path.abspath(__file__))