+++ core/__init__.py
@@ -1,11 +1,13 @@
+from config import *
+
 # core
-from core.core import *
+from .core import *
 
 # modules
-from core.state_machine import *
-from core.shader import *
-from core.math import *
-from core.tilemap import *
-from core.agent import *
-from core.animation import *
-from core.a_star import *
+from .state_machine import *
+from .shader import *
+from .math import *
+from .tilemap import *
+from .agent import *
+from .sprite import *
+from .a_star import *