+++ preview.py
@@ -10,9 +10,14 @@ class PreviewModule(Engine):
         # self.core.instantiate(Projection)
         # self.core.instantiate(AiTownSpawner)
         # self.core.instantiate(AiSimulationSpawner)
-        self.core.instantiate(Spaceship)
+        # self.core.instantiate(Spaceship)
         # self.core.instantiate(GridView, grid_size=self.core.window_size)
-
-
+        # self.core.instantiate(ScreenBlink)
+        # self.core.instantiate(DrawSystem)
+        # self.core.instantiate(DrawSystemImage)
+        # self.core.instantiate(DrawSystemAnimationImage)
+        # self.core.instantiate(DrawSystemAnimationImageSliced)
+        # self.core.instantiate(DrawSystemAnimationImageSlicedWithPaddings)
+        self.core.instantiate(DrawSystemDirectoryAnimationImageSliced)
 
 Core(background_color=(255, 255, 255, 0), fps=120)