+++ docs/reference/changelog.rst
@@ -4,10 +4,42 @@ Changelog
 v1.5
 ^^^^
 
+v1.5.1 - v1.5.2
+---------------
+
+**Documentation and structure update**
+
+* Added missing documentation informations of core.py
+* Added documentation code copy button
+* Added screen_blink.py
+* Changed Coroutine Docs, examples added
+* Explanation of Surface and Sprites
+* Changed SpriteAnimator documentation, SpriteAnimator and Simple sprites are now sections of ``Drawing System with Animations``
+* Added new parameter ``surface`` to ``TileMapDrawer.on_enable``
+* Added tutorial documentation for draw_system.py
+* Changed animation.py -> sprite.py
+* Docutils added to ``spaceship.py``
+* Gfx/Image handler added:
+    * Slicer
+    * SizeSlicer
+    * AmountSlicer
+    * SimpleSprite
+    * SimpleImageSprite
+    * SimpleSpriteAnimator
+    * SpriteDirectoryAnimation
+* Example code added:
+    * DrawSystem
+    * DrawSystemImage
+    * DrawSystemAnimationImage
+    * DrawSystemAnimationImageSliced
+    * DrawSystemAnimationImageSlicedWithPaddings
+    * DrawSystemDirectoryAnimationImageSliced
+* Example sprites added in ``examples/assets/pixel-adventure/``. Creator: https://pixelfrog-assets.itch.io/pixel-adventure-1
+
 v1.5 - v1.5.1
 -------------
 
-Documentation update
+**Documentation update**
 
 * Added ``Basic``  doc page
 * Bug Fix google meta tag